==> Part 1- Movie.ipynb <==
#### Import Data

#Imports
import pandas as pd
import numpy as np

# import basics dataset and convert to cvs
basics = pd.read_csv("Data/title.basics_real.tsv.gz", sep='\t', low_memory=False)
basics.head()

# import akas dataset and convert to cvs
akas = pd.read_csv('Data/title.akas_real.csv', low_memory=False)
akas.head()

# import rating dataset and convert to cvs
rating = pd.read_csv("Data/title.ratings_real.tsv.gz", sep='\t', low_memory=False)
rating.head()

### Replace /N with NaN

#Replace "\N" with np.nan
akas = akas.replace({'\\N':np.nan})
#Verify the change
akas.head()

akas.info()

#Check to make sure the US is the only value in region
akas['region'].value_counts()

#Replace "\N" with np.nan
basics = basics.replace({'\\N':np.nan})
#Verify the change
basics.head()

basics.info()

#Replace "\N" with np.nan
rating = rating.replace({'\\N':np.nan})
#Verify the change
rating.head()

rating.info()

### Basics

- Keep only US movies

- Eliminate movies that are null for runtimeMinutes

- Eliminate movies that are null for genre

- Keep only titleType==Movie

- Convert the startYear column to float data type.

- Filter the dataframe using startYear. Keep years between 2000-2021 (Including 2000 and 2021)

- Eliminate movies that include "Documentary" in the genre

# Filter the basics table down to only include the US by using the filter akas dataframe
keepers =basics['tconst'].isin(akas['titleId'])
keepers
#implament filter
basics = basics[keepers]
#verify changes
basics.info()

basics.head()

# only include full lenght movies 
basics = basics.loc[basics['titleType']== "movie"]
basics.info()

# drop movies with NA in runtime minutes or genres
basics = basics.dropna(subset = ['runtimeMinutes', 'genres'])
basics.info()

# coerce all to numeric to be able to filter out the years by numbers
basics['startYear']= pd.to_numeric(basics['startYear'], errors= 'coerce')

#Filter the dataframe using startYear. 
#Keep years between 2000-2021 (Including 2000 and 2021)
basics = basics[(basics['startYear'] >= 2000) & (basics['startYear'] <= 2021)]
#Verify the change with .info
basics.info()

# Exclude movies that are included in the documentary category.

is_documentary = basics['genres'].str.contains('Documentary',case=False)
basics = basics[~is_documentary]
basics.info()

#Eliminate movies that include "Documentary" in the genre
is_documentary = basics['genres'].str.contains('documentary',case=False)
basics = basics[~is_documentary]
#Verify the changes with .info
basics.info()

### Rating
Keep only movies that were included in your final title basics dataframe above


#Keep only movies that were included in your final title basics dataframe
keepers = rating['tconst'].isin(basics['tconst'])
keepers

#Filter basics
rating = rating[keepers]
#Verify the change
rating.info()

### Save data in new file

#Save basics as csv.gz file
basics.to_csv("Data/title_basics_final.csv.gz",compression='gzip',index=False)

#Open saved file and preview again
basics = pd.read_csv("Data/title_basics_final.csv.gz", low_memory = False)
basics.head()

#Save rating to a compressed file
rating.to_csv("Data/title_ratings_final.csv.gz",compression='gzip',index=False)

#Open saved file
basics = pd.read_csv("Data/title_ratings_final.csv.gz", low_memory = False)
basics.head()

#Save akas as csv.gz file
akas.to_csv("Data/title_akas_final.csv.gz",compression='gzip',index=False)
akas.head()
