==> Part 3- Databases.ipynb <==
### Part 3
9/7/23
Kass Doran 

Your stakeholder wants you to take the data you have been cleaning and collecting in Parts 1 & 2 of the project and wants you to create a MySQL database for them.

- Normalize the tables before adding them to your new database
- Keep TMDB API in 1 table together 
- Only need imdb_id, revenue, budget, and certification columns

They want the data from the following files included in your database:

Title Basics:
- Movie ID (tconst)
- Primary Title
- Start Year
- Runtime (in Minutes)
- Genres (split into two columns)
    - genre_id
    - genre_name
- title_genres (split into two columns)
    - tconst
    - genre_id
- drop the following columns:
    - "original_title" 
    - "isAdult" 
    - "titleType" 
    - "genres" and other variants of genre

Title Ratings
 - Movie ID (tconst)
 - Average Movie Rating
 - Number of Votes

The TMDB API Results (multiple files)
- Movie ID
- Revenue
- Budget
- Certification (MPAA Rating)

### Import Data

#Imports
import pandas as pd
from sqlalchemy import create_engine
from sqlalchemy_utils import database_exists, create_database
import pymysql
pymysql.install_as_MySQLdb()
from urllib.parse import quote_plus as urlquote

##log into my sql

##Load in data sets
#Title_basics
title_basics = pd.read_csv('Data/title_basics_final.csv.gz')
title_basics.head()

##Load in data sets
#Title_ratings
title_ratings = pd.read_csv('Data/title_ratings_final.csv.gz')
title_ratings.head()

##Load in data sets
#TMDB_2000 movies
df_2000 = pd.read_csv('Data/final_tmdb_data_2000.csv.gz')
df_2000.head()

##Load in data sets
#TMDB_2001 movies
df_2001 = pd.read_csv('Data/final_tmdb_data_2001.csv.gz')
df_2001.head()

#Concatenate TMDB files into one dataframe
tmdb_df = pd.concat((df_2000, df_2001))
tmdb_df.info()

# Create connection string using credentials following this format
# connection = "dialect+driver://username:password@host:port/database"
username = "root"
password = "Kassie18*" # (whatever password you chose during mysql installation)
db_name = "Project3-3"
connection = f"mysql+pymysql://{username}:{password}@localhost/{db_name}"
engine = create_engine(connection)

## Check if database exists, if not, create it
if database_exists(connection):
    print('It exists!')
else:
    create_database(connection)
    print('Database created!')

### table Normalization

#Drop columns from tmbd_data
tmdb_df = tmdb_df.drop(columns = ['adult', 'backdrop_path', 
                                      'belongs_to_collection', 'genres', 
                                      'homepage', 'id', 'original_language', 
                                      'original_title', 'overview', 'popularity', 
                                      'poster_path', 'production_companies', 
                                      'production_countries', 'release_date', 
                                      'runtime', 'spoken_languages', 'status', 
                                      'tagline', 'title', 'video', 'vote_average', 
                                      'vote_count'])
#Verify the drop
tmdb_df.info()

### Title Basics

Title Basics:

- Movie ID (tconst)
- Primary Title
- Start Year
- Runtime (in Minutes)


- Genres (split into two columns)
  - genre_id
  - genre_name
  
  
title_genres (split into two columns)
  - tconst
  - genre_id
  
  
drop the following columns:

- "original_title"
- "isAdult"
-  "titleType"
- "genres" and other variants of genre


title_basics.info()

## create a col with a list of genres
title_basics['genres_split'] = title_basics['genres'].str.split(',')
title_basics


exploded_genres = title_basics.explode('genres_split')
exploded_genres

unique_genres = sorted(exploded_genres['genres_split'].unique())
unique_genres

#Create title_genres table
title_genres = exploded_genres[['tconst', 'genres_split']].copy()
title_genres.head()

#Create genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

#Make new integer column in title_genres table
title_genres['genre_id'] = title_genres['genres_split'].map(genre_map)
title_genres = title_genres.drop(columns = 'genres_split')
title_genres.head()

#Convert genre map dictionary into genres table
genres = pd.DataFrame({'Genre_Name': genre_map.keys(),
                      'Genre_ID': genre_map.values()})
genres.head()

### Add Basics, Genere and Rating to SQL

#Drop columns from title_basics df
title_basics = title_basics.drop(columns = ['titleType', 'originalTitle', 
                                            'isAdult', 'endYear', 'genres', 
                                            'genres_split'])
#Verify the change
title_basics.info()

from sqlalchemy.types import *

#Calculate max string lengths for object columns
id_len = title_basics['tconst'].fillna('').map(len).max()
title_len = title_basics['primaryTitle'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
basics_schema = {
    "tconst": String(id_len+1), 
    "primaryTitle": Text(title_len+1),
    'startYear':Float(),
    'runtimeMinutes':Integer()}

#Save table to sql
title_basics.to_sql('title_basics', engine, dtype = basics_schema, 
                    if_exists = 'replace', index = False)

#Save table to sql
title_genres.to_sql('title_genres', engine, if_exists = 'replace')

#Save table to sql
genres.set_index('Genre_ID').to_sql('genres', engine, if_exists = 'replace', index=True)


## Title Ratings
- Movie ID (tconst)
- Average Movie Rating
- Number of Votes

#Calculate max string lengths for object columns
id_len = title_ratings['tconst'].fillna('').map(len).max()

## Create a schema dictonary using Sqlalchemy datatype objects
ratings_schema = {
    "tconst": String(id_len+1), 
    'averageRating':Float(),
    'numVotes':Integer()}

#Save table to sql
title_ratings.to_sql('title_ratings', engine, dtype = ratings_schema, 
                    if_exists = 'replace', index = False)

#Add tconst as primary key
engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')

### Tmdb Data

#Check data types of df
tmdb_df.dtypes

#Drop first row
tmdb_df = tmdb_df.drop([0])
tmdb_df.head()

#Calculate max string lengths for object columns
id_len = tmdb_df['imdb_id'].fillna('').map(len).max()
cert_len = tmdb_df['certification'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
tmdb_schema = {
    "imdb_id": String(id_len+1),
    'budget':Float(),
    'revenue':Float(),
    'certification':Text(cert_len+1)}

#Save table to sql
tmdb_df.to_sql('tmdb_data', engine, dtype = tmdb_schema, 
                    if_exists = 'replace', index = False)

#Add imdb_id as primary key
engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

#Final query
#Show tables in database
q = """SHOW TABLES;"""
pd.read_sql(q, engine)
